+++ magefile.go
@@ -123,7 +123,7 @@ func (t Test) Integration() error {
 			"github.com/aquasecurity/trivy-operator/pkg/plugin/trivy,"+
 			"github.com/aquasecurity/trivy-operator/pkg/configauditreport,"+
 			"github.com/aquasecurity/trivy-operator/pkg/vulnerabilityreport",
-		"./itest/trivy-operator")
+		"./tests/itest/trivy-operator")
 }
 
 // Target for checking if KUBECONFIG environment variable is set.
