+++ tests/itest/trivy-operator/behavior/behavior.go
@@ -7,7 +7,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/aquasecurity/trivy-operator/itest/helper"
+	"github.com/aquasecurity/trivy-operator/tests/itest/helper"
 	appsv1 "k8s.io/api/apps/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	batchv1beta1 "k8s.io/api/batch/v1beta1"
